Fix ox_thrift_client's handling of closed connections.
It turns out that when the server closes its end of the socket, the
client may not discover this until it attempts to read from the socket.
Fix this by retrying if reading the reply length returns a closed
error.

Also, don't reset the seqid when closing the socket; otherwise we need
to reencode the request with the new seqid.