New ox_thrift_socket_pool module.
ox_thrift_socket_pool maintains a pool of open sockets to a server.  A
client calls `new` to create a new pool, `destroy` to shut down a pool,
`checkout` to borrow a socket from the pool, and `checkin` to return the
socket to the pool.

This is modeled on Anthony Molinaro's ox_thrift_multi_client module, but
it moves the socket opening and closing code from the client into the
gen_server so that the process checking out the socket can be monitored,
and the socket can be released from the pool if the owning process dies.