ox_thrift_socket_pool: Open new connections in client instead of server.  Version 0.19.1.
Previously when there were no idle open sockets, the ox_thrift_socket_pool
gen_server would open a new socket and return that socket to the client.  This
would cause other clients to be blocked for the duration of the open call.
Now the server's return value tells the client to open the socket itself, and
the server can handle other requests while the open happens.