Thrift server returns exception to client when handle_function throws unknown type.
The OX Thrift server transforms "throw" calls into declared exceptions and
"error" calls into application exceptions.  Previous, it would die if it got a
"throw" with a reason that was not a tuple.  It now transforms those into
application exceptions.

Transform a throw of an unexpected type from the Thrift server's
handle_function to an "unknown" application exception.