Log an error when socket not found in ox_thrift_socket_pool dead client handling.
We are seeing a crash where ox_thrift_socket_pool is trying to close an
undefined socket that appears to happen when it receives a DOWN message from a
process for which it doesn't have a corresponding socket in the connections
map, which is something that shouldn't happen.  The only scenario that I can
think of that it would is:

  Client checks out socket.
  Client checks in socket, check message sent to pool.
  Client exits, DOWN message sent to pool.
  Pool receives checkin message, removes socket from connections.
  Pool calls demonitor(_, [ flush ]) (BUG: but DOWN message is not removed).
  Pool receives DOWN message, monitor_ref is not found.

I also added handling if a socket is not found in maybe_put_idle.  This could
happen in the following scenario:

  Client checks out socket.
  Client checks in socket, check message sent to pool.
  Client exits, DOWN message sent to pool.
  Pool receives DOWN message (BUG: out of order), removes socket from connections.
  Pool receives checkin message, socket is not found.