Fix skipping of unknown structs.  Version 0.13.0.
There was a bug that caused unknown container fields that themselves contained
containers to crash instead of being properly skipped.  For some reason that I
can no longer recall, I thought that the skip code had to call back into the
decode code, but that is not the case.

Update tests for skipping containers that contain structs.