Create improper lists when possible in ox_thrift_protocol_binary.  Version 0.14.0.
When we know that the last element of a list is a binary we can create an
improper list that is still an iolist and thus save a cons cell.

In addition when writing a byte, we can return an integer instead of a
one-byte binary and avoid the overhead of the refc_binary.