ox_thrift_server: Fix handling of oneway void functions that throw an exception.
Previously ox_thrift_server would return an exception message for any call
that threw an exception, even a call to a oneway void function.  This would
cause the next call from that client to fail with a seqid mismatch, because it
would read the exception message from the oneway void call as its reply.