ox_thrift_tests: Use an ets table for persistent data in put/get test.
This is in preparation for the spawn_options option, which will spawn a new
process to handle each request.  Previously the server used the process
dictionary to store this data, but this will no longer be persistent from one
call to the next when the spawn_options option is used.