ox_thrift_socket_pool: Do selective receive of checkin when idle queue is empty.
When a checkout request is received but the idle queue is empty,
ox_thrift_socket_pool now checks for a pending checkin request in its message
queue that can be used to handle the checkout request of opening a new socket.
This requires knowledge of gen_server's internals to for the format of
gen_server:cast messages.