ox_thrift_socket_pool: Add spares_factor option to maintain idle pool.
If a socket is closed and the number of idle sockets in the pool is below a
threshold then a new socket will be opened asynchronously in a separate
process to attempt to maintain a pool of idle sockets.

The number of idle spares to keep is a function of `spares_factor` F and the
number of Busy sockets.  If F is 0 then sockets will not be opened
preemptively.  If F is less than 1, then a new socket will be opened if the
number of Idle sockets is less than Busy ^ F; for example if there are 100
Busy sockets and F is 0.5, then a new socket will be opened if the number of
Idle sockets is less than 10 (== 100 ^ 0.5).  If F is greater than or equal to
1, then a new socket will be opened if Idle is less than Busy * F.  The
default `spares_factor` is 0.7.