rebar.config.script: Allow specification of rebar base_dir in REBAR_BASE_DIR envvar.
This can be used to speed up compilation when the source is not on a local
filesystem, for example, when the build is being done in a container with the
sources mounted from the host OS into the container.  In that case setting
REBAR_BASE_DIR to a local directory in the guest OS can speed build time.